+++ rust/src/lib.rs
@@ -8,4 +8,6 @@ pub mod authentication;
 pub mod proof;
 pub mod wns;
 
+pub const VERSION: &str = "0.8.2";
+
 pub use keys::{PrivateKeyMaterial, PublicKeyMaterial};